+++ datastore/couchDB/couchDatastoreImpl_test.go
@@ -104,4 +104,5 @@ func RunTenantServiceDatastoreTests(t *testing.T) {
 	tester.RunTenantThresholdProfileCRUD(t)
 	tester.RunGetMonitoredObjectByDomainMapTest(t)
 	tester.RunHasDashboardWithDomainTest(t)
+	tester.RunTenantReportScheduleCRUD(t)
 }