+++ datastore/druid/druidDatastoreImpl.go
@@ -466,6 +466,7 @@ func (dc *DruidDatastoreClient) GetSLAReport(request *metrics.SLAReportRequest, 
 		TimeSeriesResult:     slaTimeSeries,
 		ByHourOfDayResult:    hourOfDayBucketMap,
 		ByDayOfWeekResult:    dayOfWeekBucketMap,
+		SLAReportRequest:     *request,
 	}
 
 	data := []map[string]interface{}{}